+++ djmoney/admin.py
@@ -1,9 +1,14 @@
 # coding: utf-8
+import django.contrib.admin.helpers as admin_helpers
+import django.contrib.admin.templatetags.admin_list as admin_list
 import django.contrib.admin.utils as admin_utils
 from django import VERSION
 
 from ._compat import text_type
 from .models.fields import MoneyField
+
+
+MODULES_TO_PATCH = [admin_utils, admin_helpers, admin_list]
 
 
 def setup_admin_integration():
@@ -23,4 +28,5 @@ def setup_admin_integration():
                 return text_type(value)
             return original_display_for_field(value, field, empty)
 
-    admin_utils.display_for_field = display_for_field
+    for mod in MODULES_TO_PATCH:
+        setattr(mod, 'display_for_field', display_for_field)